pyproject.toml
[tool.poetry]
name = "bicycle-detector"
version = "0.1.0"
description = ""
authors = ["Seto Balian <seto.balian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.37.1"
torch = {version = "^2.1.2+cu118", source = "torch-cu118"}
pillow = "^10.2.0"
ffmpeg-python = "^0.2.0"
typer = {extras = ["all"], version = "^0.9.0"}
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.8.0"
black = "^23.12.1"
isort = "^5.13.2"
ipykernel = "^6.29.0"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "torch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.scripts]
bicycle-detect = 'bicycle_detector.cli:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
known_first_party = ["bicycle_detector"]


[tool.mypy]
ignore_missing_imports = true
follow_imports = "silent"
check_untyped_defs = true
enable_error_code = ["ignore-without-code"]
show_error_codes = true
warn_unused_ignores = false
